Fixing URL generation from file path.
Appending 'file:/' at the start of a file path work only under windows, where paths start with the drive letter (-> file:/c:/...).
On Unix, paths start with /, resulting in a URL like file://var/lib/... - this is then interpreted as a file on the host 'var', in directory 'lib/...'

This fix utilizes Java to source the URL for a file, instead of crafting it manually.