Extract Alerting config to common library so that it can be included in AleartBreach (#2496)
We require the ability to see the Alert configuration within the
AlertBreach. So that we have access to the config schema from the API
modules, I have moved the relevant sections from Configuration into the
AlertingConfig class in the `agent-alerting-api` module.

The main open question here is, is it acceptable that the Configuration
definition gets split up into a separate class/module.